+++ src/features/menu/hooks/useSelectMenuHook.tsx
@@ -1,17 +1,5 @@
 import { create } from 'zustand'
-
-interface OptionState {
-  selectedOptions: Record<number, number[]>
-  selectOption: (optionGroupId: number, optionId?: number) => void
-  menuStock: number
-  setMenuStock: (n: number) => void
-  price: number
-  setPrice: (n: number) => void
-  menuPrice: number
-  setMenuPrice: (n: number) => void
-  optionPrice: number
-  setOptionPrice: (n: number) => void
-}
+import { OptionState } from '../types'
 
 export const useOptionStore = create<OptionState>((set) => ({
   selectedOptions: {},
