+++ src/features/menu/types/index.ts
@@ -1,0 +1,12 @@
+export interface OptionState {
+  selectedOptions: Record<number, number[]>
+  selectOption: (optionGroupId: number, optionId?: number) => void
+  menuStock: number
+  setMenuStock: (n: number) => void
+  price: number
+  setPrice: (n: number) => void
+  menuPrice: number
+  setMenuPrice: (n: number) => void
+  optionPrice: number
+  setOptionPrice: (n: number) => void
+}
