+++ src/mocks/handlers/index.ts
@@ -4,7 +4,6 @@ import { memberHandlers } from './member'
 import { orderHandlers } from './order'
 import { reviewHandlers } from './review'
 import { storeDetailHandlers } from './storeDetail'
-import { addressHandlers } from './address'
 
 export const handlers = [
   ...reviewHandlers,
